+++ scripts/buildUtils.js
@@ -52,7 +52,10 @@ module.exports.getPackages = function getPackages() {
           {},
         ),
         ...(pkg.name === 'jest-circus' ? {'./runner': './runner.js'} : {}),
-        ...(pkg.name === 'expect' ? {'./build/utils': './build/utils.js'} : {}),
+        ...(pkg.name === 'expect' ? {
+          './build/utils': './build/utils.js',
+          './build/matchers': './build/matchers.js',
+        } : {}),
       },
       `Package ${pkg.name} does not export correct files`,
     );